For #44492: Move new '_sg_connection' property to base class
So that both the 'IODescriptorAppStore' and 'IODescriptorShotgunEntity' classes can benefit from the same fix
