+++ python/tank/descriptor/io_descriptor/shotgun_entity.py
@@ -66,7 +66,6 @@ class IODescriptorShotgunEntity(IODescriptorBase):
             optional=["project_id"]
         )
 
-        self._sg_connection = sg_connection
         self._entity_type = descriptor_dict.get("entity_type")
         self._name = descriptor_dict.get("name")
         self._field = descriptor_dict.get("field")
